CGE#1: starting to add svg icon
